Implement delete and dry-run mode, fix upload chunking, add cmd line.
I don't love the dependencies in `clap`, but it is really nice. With all
features disabled it comes out not too bad.

If you give libssh data in chunks of 256 KiB and up it breaks, EOF. I don't
know why and that doesn't seem to be documented. We use 128 KiB. I think
SFTP packets are 32 KiB anyway, so that's more than enough to fill the
pipe.

I'm happy with the 'dyn Remote' for dry-run mode, it feels like it
provides a true account of the remote operations.

Also I'm using KiB ('kibibyte' as 1024 multiples because apparently
that's most correct and unambiguous).